refactor(backend): split farm/cultures/serializers.py into a package
Split the 1117-line cultures serializers module into one module per
sub-area inside farm/cultures/serializers/: seed_rates (unit constants
and per-entry validation helpers), suppliers (SupplierSerializer,
supplier-data payload helpers, CultureSupplierDataSerializer),
seed_packages, public, cultures (CultureSerializer), imports
(import preview/apply DTOs) and seed_demand (seed-demand DTOs).

All bodies are moved verbatim; the package __init__ re-exports the
public names so the farm.cultures.serializers import path keeps working
unchanged (views, history views and tests import from there). Ruff
findings are identical to the old monolith (34 pre-existing E501,
1 B904); no new findings.
