+++ backend/farm/cultures/serializers/cultures.py
@@ -1,4 +1,4 @@
-"""DRF serializers for the cultures domain (cultures, suppliers, seeds, public library)."""
+"""Serializer for cultures (unit conversion, seed rates, supplier rows)."""
 
 from decimal import ROUND_HALF_UP, Decimal, InvalidOperation
 
@@ -25,366 +25,23 @@ from farm.models import (
 from farm.seed_units import (
     SEED_PACKAGE_UNIT_GRAMS,
     SEED_PACKAGE_UNIT_SEEDS,
-    SEED_RATE_UNIT_G_PER_LFM,
-    SEED_RATE_UNIT_G_PER_M2,
-    SEED_RATE_UNIT_SEEDS_PER_LFM,
-    SEED_RATE_UNIT_SEEDS_PER_M2,
-    SEED_RATE_UNIT_SEEDS_PER_PLANT,
     SEED_RATE_UNITS,
 )
 
-EMPTY_SEED_RATE_UNIT_VALUES = {None, '', '-'}
-
-PRE_CULTIVATION_SEED_RATE_UNITS = {
-    SEED_RATE_UNIT_G_PER_M2,
-    SEED_RATE_UNIT_G_PER_LFM,
-    SEED_RATE_UNIT_SEEDS_PER_M2,
-    SEED_RATE_UNIT_SEEDS_PER_LFM,
-    SEED_RATE_UNIT_SEEDS_PER_PLANT,
-}
-
-
-def _seed_rate_entry_error(method: str, payload: object) -> str | None:
-    """Return the validation error for one seed-rate-by-cultivation entry, if any."""
-    if not isinstance(payload, dict):
-        return 'Seed rate entries must be objects.'
-    try:
-        parsed_value = float(payload.get('value'))
-    except (TypeError, ValueError):
-        return 'Seed rate values must be numeric.'
-    if parsed_value <= 0:
-        return 'Seed rate values must be positive.'
-    unit = payload.get('unit')
-    if method == 'pre_cultivation' and unit not in PRE_CULTIVATION_SEED_RATE_UNITS:
-        return 'Pre-cultivation seed rate unit is unsupported.'
-    if method == 'direct_sowing' and unit not in SEED_RATE_UNITS:
-        return 'Direct sowing seed rate unit is unsupported.'
-    return None
-
-
-def _normalize_seed_rate_unit_value(value: object) -> str | None:
-    """Normalize supported seed rate units and legacy empty placeholders."""
-    if value in EMPTY_SEED_RATE_UNIT_VALUES:
-        return None
-    normalized_value = normalize_seed_rate_unit(value)
-    if normalized_value:
-        return normalized_value
-    raise serializers.ValidationError('Unsupported seed rate unit.')
-
-
-class SupplierSerializer(serializers.ModelSerializer):
-    created = serializers.BooleanField(read_only=True, default=False)
-    homepage_url = serializers.CharField(required=False, allow_blank=True, max_length=200)
-    
-    def get_image_file(self, obj):
-        if not obj.image_file_id:
-            return None
-        return {
-            'id': obj.image_file_id,
-            'storage_path': obj.image_file.storage_path,
-        }
-
-    def validate_allowed_domains(self, value):
-        if value is None:
-            return []
-        if not isinstance(value, list):
-            raise serializers.ValidationError('Bitte geben Sie eine Liste von Domains an.')
-        normalized = Supplier.normalize_allowed_domains(value)
-        invalid = [domain for domain in normalized if not Supplier._is_valid_domain(Supplier._normalize_domain(domain))]
-        if invalid:
-            raise serializers.ValidationError('Domains müssen gültige Hostnamen ohne Schema oder Pfad sein.')
-        return normalized
-
-    def validate_homepage_url(self, value):
-        homepage_url = (value or '').strip()
-        if not homepage_url:
-            return ''
-        if not homepage_url.startswith(('http://', 'https://')):
-            homepage_url = f'https://{homepage_url}'
-        from django.core.validators import URLValidator
-
-        try:
-            URLValidator()(homepage_url)
-        except ValidationError as exc:
-            raise serializers.ValidationError('Bitte geben Sie eine gültige URL ein.') from exc
-        return homepage_url
-
-    def validate_name(self, value):
-        from farm.utils import normalize_supplier_name
-
-        name = (value or '').strip()
-        if not name:
-            raise serializers.ValidationError('Dieses Feld ist erforderlich.')
-        project = _resolve_active_project_from_serializer(self)
-        if project is None:
-            return name
-        normalized = normalize_supplier_name(name) or ''
-        queryset = Supplier.objects.filter(project=project, name_normalized=normalized)
-        if self.instance is not None:
-            queryset = queryset.exclude(pk=self.instance.pk)
-        if queryset.exists():
-            raise serializers.ValidationError('Ein Lieferant mit diesem Namen existiert bereits.')
-        return name
-
-
-    class Meta:
-        model = Supplier
-        fields = ['id', 'name', 'homepage_url', 'slug', 'allowed_domains', 'created_at', 'updated_at', 'created']
-        read_only_fields = ['created_at', 'updated_at', 'slug']
-
-
-class SeedPackageSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = SeedPackage
-        fields = [
-            'id',
-            'culture',
-            'size_value',
-            'size_unit',
-            'evidence_text',
-            'last_seen_at',
-            'created_at',
-            'updated_at',
-        ]
-        read_only_fields = ['id', 'created_at', 'updated_at']
-        validators = []
-        extra_kwargs = {'culture': {'required': False}, 'size_unit': {'default': SeedPackage.UNIT_GRAMS}}
-
-
-
-
-
-
-    def validate(self, attrs):
-        attrs = super().validate(attrs)
-        project = _resolve_active_project_from_serializer(self)
-
-        culture = attrs.get('culture')
-        if culture is None and self.instance is not None:
-            culture = self.instance.culture
-        size_value = attrs.get('size_value')
-        size_unit = attrs.get('size_unit')
-
-        if project is not None and culture is not None and culture.project_id != project.id:
-            raise serializers.ValidationError({'culture': 'Culture does not belong to the active project.'})
-
-        if culture is None or size_value is None or size_unit is None:
-            return attrs
-
-        existing = SeedPackage.objects.filter(
-            culture=culture,
-            size_value=size_value,
-            size_unit=size_unit,
-        )
-
-        raw_initial_data = getattr(self, 'initial_data', None)
-        incoming_id = raw_initial_data.get('id') if isinstance(raw_initial_data, dict) else None
-        if incoming_id is not None:
-            try:
-                incoming_id = int(incoming_id)
-            except (TypeError, ValueError):
-                incoming_id = None
-
-        if incoming_id is not None:
-            existing = existing.exclude(pk=incoming_id)
-        elif self.instance is not None:
-            existing = existing.exclude(pk=self.instance.pk)
-        elif raw_initial_data is None:
-            # Nested serializer items in Culture updates do not reliably include initial_data.
-            # CultureSerializer handles de-duplication before replacing packages, so skip here.
-            return attrs
-
-        if existing.exists():
-            raise serializers.ValidationError('The fields culture, size_value, size_unit must make a unique set.')
-
-        return attrs
-
-
-def _has_text_value(value: object) -> bool:
-    return isinstance(value, str) and value.strip() != ''
-
-
-def _supplier_data_payload_has_supplier(row: dict[str, object]) -> bool:
-    return (
-        row.get('supplier_id') is not None
-        or row.get('supplier') is not None
-        or _has_text_value(row.get('supplier_name_input'))
-    )
-
-
-def _supplier_data_payload_has_information(row: dict[str, object]) -> bool:
-    return (
-        _has_text_value(row.get('supplier_product_name'))
-        or _has_text_value(row.get('supplier_product_url'))
-        or _has_text_value(row.get('supplier_url'))
-        or _has_text_value(row.get('notes'))
-        or _has_text_value(row.get('source_url'))
-        or row.get('germination_rate') is not None
-        or row.get('price') is not None
-        or bool(row.get('packaging_sizes'))
-    )
-
-
-def _is_empty_supplier_data_payload(row: dict[str, object]) -> bool:
-    return not _supplier_data_payload_has_supplier(row) and not _supplier_data_payload_has_information(row)
-
-
-class CultureSupplierDataSerializer(serializers.ModelSerializer):
-    supplier = SupplierSerializer(read_only=True)
-    supplier_id = serializers.PrimaryKeyRelatedField(
-        queryset=Supplier.objects.all(),
-        source='supplier',
-        write_only=True,
-        required=False,
-        allow_null=True,
-    )
-    supplier_name_input = serializers.CharField(write_only=True, required=False, allow_blank=True, allow_null=True)
-
-    class Meta:
-        model = CultureSupplierData
-        fields = [
-            'id',
-            'culture',
-            'project',
-            'supplier',
-            'supplier_id',
-            'supplier_name',
-            'supplier_name_input',
-            'supplier_url',
-            'supplier_product_name',
-            'supplier_product_url',
-            'packaging_sizes',
-            'germination_rate',
-            'price',
-            'notes',
-            'source_url',
-            'created_at',
-            'updated_at',
-        ]
-        read_only_fields = ['id', 'created_at', 'updated_at', 'project']
-        extra_kwargs = {'culture': {'required': False}}
-        validators = []
-
-    def _resolve_culture_for_validation(self, attrs):
-        culture = attrs.get('culture')
-        if culture is not None:
-            return culture
-        if self.instance is not None:
-            return self.instance.culture
-        return self.context.get('parent_culture')
-
-    def _validate_unique_culture_supplier(self, attrs):
-        culture = self._resolve_culture_for_validation(attrs)
-        supplier = attrs.get('supplier') or (self.instance.supplier if self.instance is not None else None)
-        if culture is None or supplier is None:
-            return
-
-        queryset = CultureSupplierData.objects.filter(culture=culture, supplier=supplier)
-        if self.instance is not None:
-            queryset = queryset.exclude(pk=self.instance.pk)
-        if queryset.exists():
-            raise serializers.ValidationError({
-                'supplier_id': 'supplier_data_duplicate',
-            })
-
-    def validate(self, attrs):
-        attrs = super().validate(attrs)
-        raw_initial_data = getattr(self, 'initial_data', None)
-        if isinstance(raw_initial_data, dict) and 'thousand_kernel_weight_g' in raw_initial_data:
-            raise serializers.ValidationError({
-                'thousand_kernel_weight_g': 'supplier_specific_tkg_unsupported',
-            })
-        project = _resolve_active_project_from_serializer(self)
-        culture = self._resolve_culture_for_validation(attrs)
-        supplier = attrs.get('supplier') if 'supplier' in attrs else (self.instance.supplier if self.instance is not None else None)
-        if project is not None and culture is not None and culture.project_id != project.id:
-            raise serializers.ValidationError({'culture': 'culture_project_mismatch'})
-        if project is not None and supplier is not None and supplier.project_id != project.id:
-            raise serializers.ValidationError({'supplier_id': 'supplier_project_mismatch'})
-        supplier_name_input = attrs.pop('supplier_name_input', None)
-        if not attrs.get('supplier') and supplier_name_input:
-            from farm.utils import normalize_supplier_name
-
-            project = _resolve_active_project_from_serializer(self)
-            if project is None:
-                raise serializers.ValidationError({'supplier': 'Supplier is required.'})
-
-            normalized = normalize_supplier_name(supplier_name_input) or ''
-            try:
-                with transaction.atomic():
-                    supplier, _ = Supplier.objects.get_or_create(
-                        project=project,
-                        name_normalized=normalized,
-                        defaults={
-                            'name': supplier_name_input.strip(),
-                            'homepage_url': 'https://example.invalid',
-                            'project': project,
-                        },
-                    )
-            except IntegrityError as exc:
-                supplier = Supplier.objects.filter(project=project, name_normalized=normalized).first()
-                if supplier is None:
-                    raise serializers.ValidationError({'supplier': 'Lieferant konnte nicht gespeichert werden.'}) from exc
-            attrs['supplier'] = supplier
-            attrs['supplier_name'] = supplier.name
-
-        supplier = attrs.get('supplier') if 'supplier' in attrs else (self.instance.supplier if self.instance is not None else None)
-        if supplier is None:
-            raise serializers.ValidationError({
-                'supplier_id': 'supplier_data_missing_supplier',
-            })
-
-        self._validate_unique_culture_supplier(attrs)
-        return attrs
-
-
-class PublicCultureSerializer(serializers.ModelSerializer):
-    created_by_label = serializers.SerializerMethodField()
-
-    class Meta:
-        model = PublicCulture
-        fields = [
-            'id',
-            'status',
-            'name',
-            'variety',
-            'notes',
-            'seed_supplier',
-            'supplier_name',
-            'crop_family',
-            'nutrient_demand',
-            'cultivation_types',
-            'cultivation_type',
-            'growth_duration_days',
-            'harvest_duration_days',
-            'propagation_duration_days',
-            'harvest_method',
-            'expected_yield',
-            'allow_deviation_delivery_weeks',
-            'distance_within_row_m',
-            'row_spacing_m',
-            'sowing_depth_m',
-            'seed_rate_value',
-            'seed_rate_unit',
-            'seed_rate_by_cultivation',
-            'sowing_calculation_safety_percent',
-            'thousand_kernel_weight_g',
-            'seeding_requirement',
-            'seeding_requirement_type',
-            'display_color',
-            'seed_packages',
-            'version',
-            'published_at',
-            'created_at',
-            'updated_at',
-            'created_by_label',
-            'source_project_culture',
-            'source_project',
-        ]
-        read_only_fields = fields
-
-    def get_created_by_label(self, obj: PublicCulture) -> str:
-        return obj.created_by_label
+from .seed_packages import SeedPackageSerializer
+from .seed_rates import (
+    EMPTY_SEED_RATE_UNIT_VALUES,
+    PRE_CULTIVATION_SEED_RATE_UNITS,
+    _normalize_seed_rate_unit_value,
+    _seed_rate_entry_error,
+)
+from .suppliers import (
+    CultureSupplierDataSerializer,
+    SupplierSerializer,
+    _is_empty_supplier_data_payload,
+    _supplier_data_payload_has_information,
+    _supplier_data_payload_has_supplier,
+)
 
 
 class CultureSerializer(serializers.ModelSerializer):
@@ -1051,67 +708,3 @@ class CultureSerializer(serializers.ModelSerializer):
             temp_instance.clean()
         except ValidationError as e:
             raise serializers.ValidationError(e.message_dict) from e
-
-
-class CultureImportPreviewItemSerializer(serializers.Serializer):
-    """Preview result for a single culture import item."""
-    status = serializers.ChoiceField(
-        choices=['create', 'update_candidate'],
-        help_text='Whether this culture would be created or matches an existing one'
-    )
-    matched_culture_id = serializers.IntegerField(
-        required=False,
-        allow_null=True,
-        help_text='ID of matched culture (only for update_candidate status)'
-    )
-    diff = serializers.ListField(
-        child=serializers.DictField(),
-        required=False,
-        help_text='List of fields that would change (only for update_candidate)'
-    )
-    import_data = serializers.DictField(
-        help_text='The culture data that would be imported'
-    )
-
-
-class CultureImportApplySummarySerializer(serializers.Serializer):
-    """Summary of a culture import apply operation."""
-    created_count = serializers.IntegerField(help_text='Number of cultures created')
-    updated_count = serializers.IntegerField(help_text='Number of cultures updated')
-    skipped_count = serializers.IntegerField(help_text='Number of cultures skipped')
-    errors = serializers.ListField(
-        child=serializers.DictField(),
-        help_text='List of errors encountered during import'
-    )
-
-
-class SeedDemandPackageSelectionSerializer(serializers.Serializer):
-    size_value = serializers.FloatField()
-    size_unit = serializers.CharField()
-    count = serializers.IntegerField()
-
-
-class SeedDemandPackageSuggestionSerializer(serializers.Serializer):
-    selection = SeedDemandPackageSelectionSerializer(many=True)
-    total_amount = serializers.FloatField()
-    overage = serializers.FloatField()
-    pack_count = serializers.IntegerField()
-    unit = serializers.CharField(required=False)
-
-
-class SeedDemandSerializer(serializers.Serializer):
-    """Read-only serializer for aggregated seed demand per culture."""
-    culture_id = serializers.IntegerField()
-    culture_name = serializers.CharField()
-    variety = serializers.CharField(allow_blank=True, allow_null=True)
-    supplier = serializers.CharField(allow_blank=True, allow_null=True)
-    selected_supplier_id = serializers.IntegerField(allow_null=True, required=False)
-    supplier_options = serializers.ListField(child=serializers.DictField(), required=False)
-    required_amount_value = serializers.FloatField(allow_null=True)
-    required_amount_unit = serializers.CharField(allow_null=True)
-    required_amount_warning = serializers.CharField(allow_null=True, required=False)
-    total_grams = serializers.FloatField(allow_null=True)
-    seed_packages = serializers.ListField(child=serializers.DictField(), required=False)
-    package_suggestion = SeedDemandPackageSuggestionSerializer(allow_null=True, required=False)
-    packages_needed = serializers.IntegerField(allow_null=True, required=False)
-    warning = serializers.CharField(allow_null=True)
